feat: refactor user account creation and authentication schemas, add slot validation schemas, and enhance OTP generation functionality
