+++ app/routes/auth.py
@@ -25,15 +25,8 @@ from app.exceptions.authorization_exceptions import (
     IncorrectOTPException,
     RequestNewOTPException,
 )
-from app.schema.auth_validation import (
-    OTPGenerationSchema,
-    LoginWithEmailValidationSchema,
-    OTPSubmissionSchema,
-    ParkingManagerAccountCreationSchema,
-    SignUpValidationSchema,
-    EmailVerificationSchema,
-)
 from app.schema.response_schema import ApiResponse
+from app.schema.user_auth_schema import UserLoginSchema, OTPLoginSchema, EmailVerificationSchema, GenerateOTPBaseSchema
 from app.services.auth_service import AuthService
 from app.services.token_service import TokenService
 from app.utils.error_handlers.auth_error_handlers import (
@@ -57,30 +50,9 @@ auth_blp = Blueprint(
 )
 
 
-@auth_blp.route("/create-new-account")
-class CreateNewAccount(MethodView):
-
-    @auth_blp.arguments(SignUpValidationSchema)
-    @auth_blp.response(201, ApiResponse)
-    @auth_blp.doc(
-        description="Create a new user account.",
-        responses={
-            201: {"description": "User created successfully."},
-            400: {"description": "Bad Request"},
-        },
-    )
-    @jwt_required(True)
-    def post(self, sign_up_data):
-        auth_service = AuthService()
-        auth_service.create_new_user(sign_up_data)
-        return set_response(
-            201, {"code": "success", "message": "Check your email for verification."}
-        )
-
-
 @auth_blp.route("/login")
 class Login(MethodView):
-    @auth_blp.arguments(LoginWithEmailValidationSchema)
+    @auth_blp.arguments(UserLoginSchema)
     @auth_blp.response(200, ApiResponse)
     @auth_blp.doc(
         description="Login with email.",
@@ -91,8 +63,7 @@ class Login(MethodView):
     )
     @jwt_required(True)
     def post(self, login_data):
-        auth_service = AuthService()
-        auth_service.login_user(login_data)
+        AuthService.login_user(login_data)
         response = set_response(
             200, {"code": "otp_sent", "message": "OTP sent successfully."}
         )
@@ -104,7 +75,7 @@ class Login(MethodView):
 
 @auth_blp.route("/generate-otp")
 class GenerateOTP(MethodView):
-    @auth_blp.arguments(OTPGenerationSchema)
+    @auth_blp.arguments(GenerateOTPBaseSchema)
     @auth_blp.response(200, ApiResponse)
     @auth_blp.doc(
         description="Generate an OTP.",
@@ -125,7 +96,7 @@ class GenerateOTP(MethodView):
 
 @auth_blp.route("/verify-otp")
 class VerifyOTP(MethodView):
-    @auth_blp.arguments(OTPSubmissionSchema)
+    @auth_blp.arguments(OTPLoginSchema)
     @auth_blp.response(200, ApiResponse)
     @auth_blp.doc(
         description="Verify the OTP.",
@@ -227,28 +198,6 @@ class VerifyEmail(MethodView):
                 "message": "Email verified successfully.",
             },
         )
-
-
-@auth_blp.route("/owner/register")
-class ParkingManagerRegistration(MethodView):
-    """Handles parking manager registration."""
-
-    @auth_blp.arguments(ParkingManagerAccountCreationSchema)
-    @auth_blp.response(201, ApiResponse)
-    def post(self, registration_data):
-        """Register a new parking manager account."""
-        # TODO: Add service call here
-        # result = ParkingManagerService.register_manager(registration_data)
-
-        return set_response(
-            201,
-            {
-                "code": "registration_success",
-                "message": "Parking manager registration successful",
-                # "data": result
-            },
-        )
-
 
 auth_blp.register_error_handler(BannedUserException, handle_banned_user)
 auth_blp.register_error_handler(EmailNotFoundException, handle_email_not_found)
