+++ app/routes/user.py
@@ -7,11 +7,9 @@ from flask_jwt_extended import jwt_required
 from flask_smorest import Blueprint
 
 from app.exceptions.authorization_exceptions import EmailAlreadyTaken
-from app.schema.auth_validation import EmailVerificationSchema
 from app.schema.response_schema import ApiResponse
 from app.schema.user_auth_schema import UserRegistrationSchema
 from app.services.auth_service import AuthService
-from app.services.user_auth import UserAuth
 from app.utils.error_handlers.auth_error_handlers import handle_email_already_taken
 from app.utils.response_util import set_response
 
@@ -37,11 +35,9 @@ class CreateUserAccount(MethodView):
     )
     @jwt_required(True)
     def post(self, sign_up_data: dict):
-        sign_up_data.update({"role": "user"})
-        UserAuth.create_new_user(sign_up_data)
+        AuthService.create_new_user(sign_up_data)
         return set_response(
             201, {"code": "success", "message": "Check your email for verification."}
         )
-
 
 user_auth_blp.register_error_handler(EmailAlreadyTaken, handle_email_already_taken)
