+++ app/schema/user_auth_schema.py
@@ -2,39 +2,23 @@
 
 from marshmallow import fields, validate, Schema, post_load
 
+from app.schema.common_registration_schema import CommonRegistrationSchema, EmailBaseSchema
 
-class UserRegistrationSchema(Schema):
+
+class UserRegistrationSchema(CommonRegistrationSchema):
     """Schema for user registration."""
-
-    email = fields.Email(required=True, validate=validate.Length(min=3, max=75))
-    first_name = fields.Str(required=True, validate=validate.Length(min=3, max=50))
-    middle_name = fields.Str(
-        required=False, missing=None, validate=validate.Length(min=0, max=50)
-    )
-    last_name = fields.Str(required=True, validate=validate.Length(min=3, max=100))
-    phone_number = fields.Str(
-        required=True,
-        validate=[
-            validate.Length(min=10, max=15),
-            validate.Regexp(
-                regex=r"^\+?[0-9]\d{1,14}$", error="Invalid phone number format."
-            ),
-        ],
-    )
     nickname = fields.Str(required=True, validate=validate.Length(min=3, max=24))
-    suffix = fields.Str(required=False, missing=None)
-
     plate_number = fields.Str(
         required=True,
         validate=[
             validate.Length(min=6, max=8),
             validate.Regexp(
                 regex=r"^(?:"
-                r"[A-Z]{2,3}[\s-]?\d{3,4}|"
-                r"CD[\s-]?\d{4}|"
-                r"[A-Z]{3}[\s-]?\d{3}|"
-                r"\d{4}"
-                r")$",
+                      r"[A-Z]{2,3}[\s-]?\d{3,4}|"
+                      r"CD[\s-]?\d{4}|"
+                      r"[A-Z]{3}[\s-]?\d{3}|"
+                      r"\d{4}"
+                      r")$",
                 error=(
                     "Invalid plate number format. Please use one of these formats:\n"
                     "• Private vehicles: ABC 123 or ABC 1234\n"
@@ -45,55 +29,44 @@ class UserRegistrationSchema(Schema):
             ),
         ],
     )
-
+    
     @post_load
-    def normalize_plate_number(
+    def normalize_data(
         self, in_data, **kwargs
     ):  # pylint: disable=unused-argument
-        """Normalize plate number format by removing spaces and converting to uppercase"""
+        """Method to normalize the input data."""
         if "plate_number" in in_data:
             in_data["plate_number"] = in_data["plate_number"].upper().replace(" ", "")
-        return in_data
-
-    @post_load
-    def normalize_email(self, in_data, **kwargs):  # pylint: disable=unused-argument
-        """Method to convert email to lowercase."""
-        if "email" in in_data:
-            in_data["email"] = in_data["email"].lower()
-        return in_data
-
-    @post_load
-    def normalize_personal_data(
-        self, in_data, **kwargs
-    ):  # pylint: disable=unused-argument
-        """Method to capitalize the first name, middle name, last name, suffix, and nickname."""
-        if "first_name" in in_data:
-            in_data["first_name"] = in_data["first_name"].capitalize()
-        if "middle_name" in in_data and in_data["middle_name"]:
-            in_data["middle_name"] = in_data["middle_name"].capitalize()
-        if "last_name" in in_data:
-            in_data["last_name"] = in_data["last_name"].capitalize()
-        if "suffix" in in_data and in_data["suffix"]:
-            in_data["suffix"] = in_data["suffix"].capitalize()
         if "nickname" in in_data:
             in_data["nickname"] = in_data["nickname"].capitalize()
         return in_data
-
-
-class UserLoginSchema(Schema):
-    """Schema for user login."""
-
-    email = fields.Email(required=True, validate=validate.Length(min=3, max=75))
-
-    @post_load
-    def normalize_email(self, in_data, **kwargs):  # pylint: disable=unused-argument
-        """Method to convert email to lowercase."""
-        in_data["email"] = in_data["email"].lower()
+    
+    @post_load()
+    def add_role(self, in_data, **kwargs):  # pylint: disable=unused-argument
+        """Method to add role to the user."""
+        in_data["role"] = "user"
         return in_data
 
 
-class OTPLoginSchema(Schema):
+class UserLoginSchema(EmailBaseSchema):
+    """Schema for user login."""
+    role = fields.Str(required=True, validate=[validate.Length(min=3, max=50), validate.OneOf(["user", "admin", "parking_manager"])])
+
+
+class OTPLoginBaseSchema(EmailBaseSchema):
+    """Schema for OTP login."""
+    otp = fields.Str(required=True, validate=validate.Length(equal=6))
+    remember_me = fields.Bool(required=False, missing=False)
+
+
+class GenerateOTPBaseSchema(EmailBaseSchema):
+    """Schema for generating OTP."""
+
+
+class OTPLoginSchema(OTPLoginBaseSchema):
     """Schema for OTP login."""
 
-    email = fields.Email(required=True, validate=validate.Length(min=3, max=75))
-    otp = fields.Str(required=True, validate=validate.Length(equal=6))
+
+class EmailVerificationSchema(Schema):
+    """Class to handle email verification validation."""
+    verification_token = fields.Str(required=True, validate=validate.Length(min=1, max=200))
